+++ src/App.css
@@ -292,11 +292,6 @@ div.media.discover div.filter_panel h2 div.media.discover div.filter_panel h3 {
   padding: 0;
 }
 
-div.media.discover div.filter_panel div.name span.chevron-right {
-  transform: rotate(90deg);
- 
-}
-
 
 .glyphicons_v2.chevron-right {
   background-image: url(/src/media/rotate.svg); 
@@ -659,9 +654,36 @@ footer nav ul li {
 } 
 
 
-/* @media (max-width: 1550px) {
-  .container {
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr 1fr 1fr;
-  }
-} */
+div.pagination.infinite {
+  width: 100%;
+  height: 50px;
+}
+
+div.pagination {
+ 
+  color: #000;
+  padding: 0;
+  margin: 0;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  align-items: center;
+  align-content: center;
+  width: auto;
+  
+}
+
+div.column_wrapper div.content_wrapper p:last-child {
+  margin-bottom: 0;
+}
+
+div.pagination p.load_more {
+  justify-content: center;
+}
+
+div.pagination.infinite p {
+  width: 100%;
+  text-align: center;
+  margin: 0;
+  padding: 0;
+}
