+++ src/App.js
@@ -1,11 +1,33 @@
 import "./App.css";
 import Header from "./components/Header";
+import Main from "./components/Main";
+import { useEffect, useState } from "react";
 
 function App() {
+  const [isVisible, setIsVisible] = useState(true);
+  const [height, setHeight] = useState(0);
+
+  useEffect(() => {
+    window.addEventListener("scroll", listenToScroll);
+    return () => window.removeEventListener("scroll", listenToScroll);
+  }, []);
+
+  const listenToScroll = () => {
+    let heightToHideFrom = 200;
+    const winScroll =
+      document.body.scrollTop || document.documentElement.scrollTop;
+    setHeight(winScroll);
+
+    if (winScroll > heightToHideFrom) {
+      isVisible && setIsVisible(false);
+    } else {
+      setIsVisible(true);
+    }
+  };
   return (
     <div>
-      <Header />
-      <div></div>
+      {isVisible && <Header />}
+      <Main />
     </div>
   );
 }
