+++ src/components/MovieCard.js
@@ -16,7 +16,6 @@ const MovieCard = ({ movie, name }) => {
   return (
     <div>
       <div className="white_column no_pad">
-        {/*Messege for future. There might be problem in css */}
         <section id="media_results" className="panel results">
           <div className="media_item results">
             <div id="page_1" className="page_wrapper">
@@ -35,18 +34,6 @@ const MovieCard = ({ movie, name }) => {
                     fill={movie.vote_average * 10}
                     rate={movie.vote_average * 10}
                   />
-                  {/* <div className="consensus tight">
-                    <div className="outer_ring">
-                      <div className="user_score_chart">
-                        <div className="percent">
-                          <span className="icon icon-r78">
-                            {}
-                          </span>
-                        </div>
-                       <Circle />}
-                      </div>
-                    </div>
-                  </div> */}
 
                   <h2>
                     <a href="#">{movie.title}</a>
@@ -69,31 +56,3 @@ const MovieCard = ({ movie, name }) => {
 };
 export default MovieCard;
 
-{
-  /* <img src ={`${IMAGE_PATH}${movie.poster_path}`} alt=""/>
-        <h5>{movie.title}</h5>
-
-
-        // {movie.poster_path ? <img src ={`${IMAGE_PATH}${movie.poster_path}`} alt=""/> : null}
-// {movie.title}
-
-
-
-
- {/* <div className='white_column no_pad'>  {/*Messege for future. There might be problem in css 
-       <section id="media_results" className="panel results">
-          <div className='media_item results'>
-            <div id='page_1' className="page_wrapper">
-                <div className="card style_1">
-                  <div className="image"></div>
-                      <div className="wrapper">
-                      <a className='image'> 
-                        <img  className='poster' src ={`${IMAGE_PATH}${movie.poster_path}`}/> 
-                      </a>
-                      </div>
-                </div>
-            </div>
-          </div>
-       </section>
-     </div> */
-}
